backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
python-dotenv==1.0.0
feedparser==6.0.10
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dateutil==2.8.2
apscheduler==3.10.4
